Extract the last actions
