--- app/src/actions.js
@@ -1,51 +1,0 @@
-// map
-export const INIT_GOOGLE_MAPS = 'INIT_GOOGLE_MAPS';
-export const SET_LAYERS = 'SET_LAYERS';
-export const SET_LAYER_HEADER = 'SET_LAYER_HEADER';
-export const SET_TILESET_URL = 'SET_TILESET_URL';
-export const SET_TILESET_ID = 'SET_TILESET_ID';
-export const SET_MAX_ZOOM = 'SET_MAX_ZOOM';
-export const SET_LAYER_OPACITY = 'SET_LAYER_OPACITY';
-export const SET_ZOOM = 'SET_ZOOM';
-export const SET_CENTER = 'SET_CENTER';
-export const SET_CENTER_TILE = 'SET_CENTER_TILE';
-export const SET_URL_WORKSPACE_ID = 'SET_URL_WORKSPACE_ID';
-export const SET_WORKSPACE_ID = 'SET_WORKSPACE_ID';
-export const DELETE_WORKSPACE_ID = 'DELETE_WORKSPACE_ID';
-export const ADD_CUSTOM_LAYER = 'ADD_CUSTOM_LAYER';
-export const SET_LOADING = 'SET_LOADING';
-export const SET_LOADERS = 'SET_LOADERS';
-export const SET_DRAWING = 'SET_DRAWING';
-export const SET_MOUSE_LAT_LONG = 'SET_MOUSE_LAT_LONG';
-
-
-// modals
-export const SET_SUPPORT_MODAL_VISIBILITY = 'SET_SUPPORT_MODAL_VISIBILITY';
-export const SET_LAYER_MANAGEMENT_MODAL_VISIBILITY = 'SET_LAYER_MANAGEMENT_MODAL_VISIBILITY';
-
-
-// Workspace layer management
-export const TOGGLE_LAYER_PANEL_EDIT_MODE = 'TOGGLE_LAYER_PANEL_EDIT_MODE';
-export const TOGGLE_LAYER_VISIBILITY = 'TOGGLE_LAYER_VISIBILITY';
-export const TOGGLE_LAYER_WORKSPACE_PRESENCE = 'TOGGLE_LAYER_WORKSPACE_PRESENCE';
-export const SET_LAYER_HUE = 'SET_LAYER_HUE';
-export const SET_WORKSPACE_LAYER_LABEL = 'SET_WORKSPACE_LAYER_LABEL';
-export const SHOW_CONFIRM_LAYER_REMOVAL_MESSAGE = 'SHOW_CONFIRM_LAYER_REMOVAL_MESSAGE';
-
-// Appearance actions
-export const SET_LAYER_INFO_MODAL = 'SET_LAYER_INFO_MODAL';
-
-// Vessel info actions
-export const ADD_VESSEL = 'ADD_VESSEL';
-export const SET_VESSEL_DETAILS = 'SET_VESSEL_DETAILS';
-export const SHOW_VESSEL_DETAILS = 'SHOW_VESSEL_DETAILS';
-export const SET_VESSEL_TRACK = 'SET_VESSEL_TRACK';
-export const CLEAR_VESSEL_INFO = 'CLEAR_VESSEL_INFO';
-export const HIDE_VESSELS_INFO_PANEL = 'HIDE_VESSELS_INFO_PANEL';
-export const SET_TRACK_BOUNDS = 'SET_TRACK_BOUNDS';
-export const TOGGLE_VESSEL_PIN = 'TOGGLE_VESSEL_PIN';
-export const SET_PINNED_VESSEL_HUE = 'SET_PINNED_VESSEL_HUE';
-export const SET_PINNED_VESSEL_TITLE = 'SET_PINNED_VESSEL_TITLE';
-export const LOAD_PINNED_VESSEL = 'LOAD_PINNED_VESSEL';
-export const TOGGLE_PINNED_VESSEL_EDIT_MODE = 'TOGGLE_PINNED_VESSEL_EDIT_MODE';
-export const SET_PINNED_VESSEL_TRACK_VISIBILITY = 'SET_PINNED_VESSEL_TRACK_VISIBILITY';
