+++ app/src/actions/map.js
@@ -1,20 +1,19 @@
-import {
-  INIT_GOOGLE_MAPS,
-  SET_ZOOM,
-  SET_CENTER,
-  SET_LOADING,
-  SET_DRAWING,
-  SET_LOADERS,
-  DELETE_WORKSPACE_ID,
-  SET_LAYER_INFO_MODAL,
-  SET_SUPPORT_MODAL_VISIBILITY,
-  SET_LAYER_MANAGEMENT_MODAL_VISIBILITY,
-  SET_CENTER_TILE,
-  SET_MOUSE_LAT_LONG
-} from 'actions';
 import { clearVesselInfo } from 'actions/vesselInfo';
 import { trackCenterTile } from 'analytics/analyticsActions';
 import { ANALYTICS_TILE_COORDS_SCALE, ANALYTICS_TRACK_DRAG_FROM_ZOOM, CLUSTER_CLICK_ZOOM_INCREMENT } from 'constants';
+
+export const SET_LAYER_MANAGEMENT_MODAL_VISIBILITY = 'SET_LAYER_MANAGEMENT_MODAL_VISIBILITY';
+export const INIT_GOOGLE_MAPS = 'INIT_GOOGLE_MAPS';
+export const SET_ZOOM = 'SET_ZOOM';
+export const SET_CENTER = 'SET_CENTER';
+export const SET_LOADING = 'SET_LOADING';
+export const SET_LOADERS = 'SET_LOADERS';
+export const SET_DRAWING = 'SET_DRAWING';
+export const SET_MOUSE_LAT_LONG = 'SET_MOUSE_LAT_LONG';
+export const SET_CENTER_TILE = 'SET_CENTER_TILE';
+export const SET_LAYER_INFO_MODAL = 'SET_LAYER_INFO_MODAL';
+export const DELETE_WORKSPACE_ID = 'DELETE_WORKSPACE_ID';
+
 
 // store the original google maps in the app state.
 // this is needed in the heatmap actions/reducers, to avoid constantly passing
@@ -142,13 +141,6 @@ export function openTimebarInfoModal() {
 export function zoomIntoVesselCenter(latLng) {
   return (dispatch, getState) => {
     dispatch(setZoom(getState().map.zoom + CLUSTER_CLICK_ZOOM_INCREMENT, [latLng.lat(), latLng.lng()]));
-  };
-}
-
-export function setSupportModalVisibility(visibility) {
-  return {
-    type: SET_SUPPORT_MODAL_VISIBILITY,
-    payload: visibility
   };
 }
