+++ app/src/reducers/vesselInfo.js
@@ -1,22 +1,22 @@
 import cloneDeep from 'lodash/cloneDeep';
 import find from 'lodash/find';
 import getVesselName from 'util/getVesselName';
+import { HEATMAP_TRACK_HIGHLIGHT_HUE, VESSEL_INFO_STATUS } from 'constants';
 import {
+  SET_VESSEL_DETAILS,
   ADD_VESSEL,
+  SET_VESSEL_TRACK,
+  LOAD_PINNED_VESSEL,
+  SHOW_VESSEL_DETAILS,
   CLEAR_VESSEL_INFO,
   HIDE_VESSELS_INFO_PANEL,
-  LOAD_PINNED_VESSEL,
-  SET_PINNED_VESSEL_HUE,
-  SET_PINNED_VESSEL_TITLE,
-  SET_PINNED_VESSEL_TRACK_VISIBILITY,
+  TOGGLE_VESSEL_PIN,
   SET_TRACK_BOUNDS,
-  SET_VESSEL_DETAILS,
-  SET_VESSEL_TRACK,
-  SHOW_VESSEL_DETAILS,
-  TOGGLE_PINNED_VESSEL_EDIT_MODE,
-  TOGGLE_VESSEL_PIN
-} from 'actions';
-import { HEATMAP_TRACK_HIGHLIGHT_HUE, VESSEL_INFO_STATUS } from 'constants';
+  SET_PINNED_VESSEL_HUE,
+  SET_PINNED_VESSEL_TRACK_VISIBILITY,
+  SET_PINNED_VESSEL_TITLE,
+  TOGGLE_PINNED_VESSEL_EDIT_MODE
+} from 'actions/vesselInfo';
 
 const initialState = {
   vessels: [],
