+++ app/src/siteNav/containers/MenuMobile.js
@@ -1,8 +1,8 @@
 import { connect } from 'react-redux';
 import MenuMobile from 'siteNav/components/MenuMobile';
 import { login, logout } from 'user/userActions';
-import { setSupportModalVisibility } from 'actions/map';
 import { getWorkspace } from 'actions/workspace';
+import { setSupportModalVisibility } from 'siteNav/supportFormActions';
 
 const mapStateToProps = state => ({
   loggedUser: state.user.loggedUser
