+++ app/src/siteNav/supportFormReducer.js
@@ -1,0 +1,17 @@
+import { FORM_RESPONSE, SET_SUPPORT_MODAL_VISIBILITY } from 'siteNav/supportFormActions';
+
+const initialState = {
+  open: false,
+  supportRequestStatus: null
+};
+
+export default function (state = initialState, action) {
+  switch (action.type) {
+    case FORM_RESPONSE:
+      return Object.assign({}, state, { supportRequestStatus: action.payload });
+    case SET_SUPPORT_MODAL_VISIBILITY:
+      return Object.assign({}, state, { open: action.payload });
+    default:
+      return state;
+  }
+}
